+++ velox/experimental/cudf/connectors/hive/CudfSplitReaderHelpers.cpp
@@ -15,7 +15,7 @@
  */
 
 #include "velox/experimental/cudf/CudfNoDefaults.h"
-#include "velox/experimental/cudf/connectors/hive/CudfHiveDataSourceHelpers.h"
+#include "velox/experimental/cudf/connectors/hive/CudfSplitReaderHelpers.h"
 
 #include "velox/dwio/common/BufferedInput.h"
